+++ classes/task/cleargame.php
@@ -14,19 +14,43 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-
 /**
- * Game block version identification
+ *  Block Game config task definition class
  *
  * @package    block_game
  * @copyright  2019 Jose Wilson
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+namespace block_game\task;
 
-defined('MOODLE_INTERNAL') || die();
+/**
+ * Task to clear course deleted game
+ *
+ * @package    block_game
+ * @copyright  2019 Jose Wilson
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class cleargame extends \core\task\scheduled_task {
 
-$plugin->version  = 2022042722;
-$plugin->requires = 2018051700;
-$plugin->maturity = MATURITY_STABLE;
-$plugin->release  = '1.31 Version for Moodle 3.5 onwards';
-$plugin->component = 'block_game';
+    /**
+     * Task to clear deleted users and courses from the block game.
+     *
+     * @return string
+     *
+     * @throws \coding_exception
+     */
+    public function get_name() {
+        return get_string('cleargametask', 'block_game');
+    }
+
+    /**
+     * Execute the task.
+     */
+    public function execute() {
+        global $CFG;
+        require_once($CFG->dirroot . '/blocks/game/lib.php');
+        block_game_clear_course_deleted_game();
+        block_game_clear_user_deleted_game();
+        return true;
+    }
+}