+++ edit_form.php
@@ -102,15 +102,15 @@ class block_game_edit_form extends block_edit_form {
             $mform->setDefault('config_show_level', 1);
             $mform->addHelpButton('config_show_level', 'config_level', 'block_game');
             // Options controlling level up.
-            $levels = array(4 => 4, 6 => 6, 8 => 8, 10 => 10, 12 => 12);
+            $levels = array(4 => 4, 6 => 6, 8 => 8, 10 => 10, 12 => 12, 13 => 13, 14 => 14, 15 => 15);
             $mform->addElement('select', 'config_level_number', get_string('config_level_number', 'block_game'), $levels);
             $mform->setDefault('config_level_number', 6);
             $mform->disabledIf('config_level_number', 'config_show_level', 'eq', 0);
             $mform->addHelpButton('config_level_number', 'config_level_number', 'block_game');
             $leveluppoints = array(1 => 300, 2 => 500, 3 => 1000, 4 => 2000,
                 5 => 4000, 6 => 6000, 7 => 10000, 8 => 20000,
-                9 => 30000, 10 => 50000, 11 => 70000, 12 => 100000);
-            for ($i = 1; $i <= 12; $i++) {
+                9 => 30000, 10 => 50000, 11 => 70000, 12 => 100000, 13 => 150000, 14 => 250000, 15 => 500000);
+            for ($i = 1; $i <= 15; $i++) {
                 $mform->addElement('text', 'config_level_up' . $i, get_string('config_level_up' . $i, 'block_game'));
                 $mform->setDefault('config_level_up' . $i, $leveluppoints[$i]);
                 $mform->disabledIf('config_level_up' . $i, 'config_show_level', 'eq', 0);
@@ -127,7 +127,8 @@ class block_game_edit_form extends block_edit_form {
             $mform->addElement('html', get_string('title_config_section', 'block_game'));
             $sections = block_game_get_sections_course($COURSE->id);
             foreach ($sections as $section) {
-                $limit = array(0 => 0, 5 => 5, 10 => 10, 20 => 20, 30 => 30, 50 => 50, 60 => 60, 80 => 80, 100 => 100);
+                $limit = array(0 => 0, 5 => 5, 10 => 10, 15 => 15, 20 => 20, 25 => 25, 30 => 30, 35 => 35, 40 => 40,
+                    45 => 45, 50 => 50, 60 => 60, 70 => 70, 80 => 80, 90 => 90, 100 => 100);
                 $txtsection = get_string('section', 'block_game') . ' ' . $section->section;
                 if (isset($section->name) && $section->name != "") {
                     $txtsection = $section->name;
@@ -142,7 +143,8 @@ class block_game_edit_form extends block_edit_form {
             $activities = $coursedata['activities'];
             foreach ($activities as $activity) {
                 if (block_game_is_visibled_module($COURSE->id, $activity['id'])) {
-                    $limit = array(0 => 0, 5 => 5, 10 => 10, 20 => 20, 30 => 30, 50 => 50, 60 => 60, 80 => 80, 100 => 100);
+                    $limit = array(0 => 0, 5 => 5, 10 => 10, 15 => 15, 20 => 20, 25 => 25, 30 => 30, 35 => 35, 40 => 40,
+                    45 => 45, 50 => 50, 60 => 60, 70 => 70, 80 => 80, 90 => 90, 100 => 100);
                     $attributes = ['class' => 'iconlarge activityicon'];
                     $icon = $OUTPUT->pix_icon('icon', $activity['modulename'], $activity['type'], $attributes);
                     $activityoption = array();