+++ settings.php
@@ -36,7 +36,7 @@ if ($ADMIN->fulltree) {
             new lang_string('imagemavatardesc', 'block_game'),
             'imagens_avatar',
             0,
-            ['subdirs' => 0, 'maxfiles' => 56, 'accepted_types' => array('.svg')]
+            ['subdirs' => 0, 'maxfiles' => 68, 'accepted_types' => array('.svg')]
         ));
 
     $settings->add(new admin_setting_configcheckbox('block_game/change_avatar_course',
@@ -74,18 +74,18 @@ if ($ADMIN->fulltree) {
             new lang_string('imagemlevelsdesc', 'block_game'),
             'imagens_levels',
             0,
-            ['subdirs' => 0, 'maxfiles' => 13, 'accepted_types' => array('.svg')]
+            ['subdirs' => 0, 'maxfiles' => 16, 'accepted_types' => array('.svg')]
         ));
 
     // Options controlling level up.
-    $leveloptions = array(4 => 4, 6 => 6, 8 => 8, 10 => 10, 12 => 12);
+    $leveloptions = array(4 => 4, 6 => 6, 8 => 8, 10 => 10, 12 => 12, 13 => 13, 14 => 14, 15 => 15);
     $settings->add(new admin_setting_configselect('block_game/level_number',
-            get_string('config_level_number', 'block_game'), '', 12, $leveloptions));
+            get_string('config_level_number', 'block_game'), '', 15, $leveloptions));
 
     $leveluppoints = array(1 => 300, 2 => 500, 3 => 1000, 4 => 2000,
         5 => 4000, 6 => 6000, 7 => 10000, 8 => 20000,
-        9 => 30000, 10 => 50000, 11 => 70000, 12 => 100000);
-    for ($i = 1; $i <= 12; $i++) {
+        9 => 30000, 10 => 50000, 11 => 70000, 12 => 100000, 13 => 150000, 14 => 250000, 15 => 500000);
+    for ($i = 1; $i <= 15; $i++) {
         $settings->add(new admin_setting_configtext('block_game/level_up' . $i,
                 get_string('config_level_up' . $i, 'block_game'), '', $leveluppoints[$i], PARAM_INT));
     }